T 174 fetch top playgrounds from db (#130)
* t-174-add fetch top courts from DB

* t-174-fix getServerSideProps method didn't work

* Delete Home.js
